Fix FIELD_BOOLEAN using int size instead of byte size for saving and restoring Resolves #65

+++ dlls/util.cpp
@@ -1676,7 +1676,7 @@ static int gSizes[FIELD_TYPECOUNT] =
 #else
 	sizeof(int *),		// FIELD_FUNCTION	
 #endif
-	sizeof(int),		// FIELD_BOOLEAN
+	sizeof(byte),		// FIELD_BOOLEAN
 	sizeof(short),		// FIELD_SHORT
 	sizeof(char),		// FIELD_CHARACTER
 	sizeof(float),		// FIELD_TIME
